Create a standalone test to test replication with only status available on author. Rename ReplicationIT to PublishEndToEndIT - This test is intended to test the overall flow for publishing a page, the page being accessible on publish.
